refactor(test): extract infra-check token wiring test to own module
Move TestInfraCheckWorkflowTokenWiring from test_check_agent_infrastructure.py
(which was at 518 lines, triggering a taste violation) into a dedicated
test_infra_check_workflow_token_wiring.py module. Restore taste baseline to 583.

Original file now 499 lines, under the 500-line threshold.
