+++ scripts/ci/taste_count_baseline.txt
@@ -1,1 +1,1 @@
-584
+583
